data_analysis/requirement.txt
ipykernel
numpy
openpyxl
seaborn
matplotlib
xlsxwriter
Flask
xlrd>=2.0.1
boto3
python-dotenv